fixup variable size
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/ec/trunk@38839 72102866-910b-0410-8b05-ffd578937521

+++ EcProcess.m
@@ -4620,8 +4620,8 @@ With two parameters ('maximum' and a number),\n\
     {
       NSInteger     inc;
       NSInteger     pct;
-      NSInteger     iMax = 0;
-      NSInteger     pMax = 0;
+      uint64_t      iMax = 0;
+      uint64_t      pMax = 0;
 
       /* We increase the threshold for the next alert by a percentage
        * of the existing usage or by a fixed increment, whichever is
